requirements.txt
os
json
hashlib
streamlit
langchain
langchain_core
langchain_community
angchain_google_genai
langchain_text_splitters